% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{seatable_alltables}
\alias{seatable_alltables}
\title{List all tables across all bases visible to a connection}
\usage{
seatable_alltables(con = default_connection())
}
\arguments{
\item{con}{A \link{seatable_connection}.}
}
\value{
A data.frame with \code{base_name}, \code{workspace_id}, table \code{name} and \verb{_id}.
}
\description{
List all tables across all bases visible to a connection
}
