% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{seatable_workspaces}
\alias{seatable_workspaces}
\title{List workspaces (and their bases) visible to a connection}
\usage{
seatable_workspaces(con = default_connection())
}
\arguments{
\item{con}{A \link{seatable_connection}.}
}
\value{
A data.frame with at least \code{workspace_id} and base \code{name} columns.
}
\description{
List workspaces (and their bases) visible to a connection
}
