% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{set_default_connection}
\alias{set_default_connection}
\title{Set (or clear) the default connection}
\usage{
set_default_connection(con)
}
\arguments{
\item{con}{A \link{seatable_connection}, or \code{NULL} to clear the default.}
}
\value{
The previous default connection, invisibly (or \code{NULL}).
}
\description{
Set (or clear) the default connection
}
