% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{with_connection}
\alias{with_connection}
\title{Evaluate an expression with a temporary default connection}
\usage{
with_connection(con, expr)
}
\arguments{
\item{con}{A \link{seatable_connection} to use as the default for the duration of
\code{expr}.}

\item{expr}{Expression to evaluate.}
}
\value{
The value of \code{expr}.
}
\description{
Evaluate an expression with a temporary default connection
}
\examples{
\dontrun{
with_connection(seatable_connection("https://cloud.seatable.io/"), {
  seatable_query("SELECT * FROM banc_meta")
})
}
}
